TT-12519, CR feedback
