+++ tests/proxy/proxy.go
@@ -1,0 +1,1 @@
+package proxy
